feat(DocSearch): search as you type
Algolia offers DocSearch for free without any limits on the amount of operations, so offering search as you type is actually a better experience.

Because updating the URL on every keypress is actually a more expensive operation than actually doing the search, I delayed that by 700ms after the last keypress.

Meanwhile I also updated the Algolia logo to be the newer logo.

Something negative about this change could be that the 'loading' bar now goes at every keypress, Is there some way to delay that, or maybe disable all together for the dispatching of SearchDone? I couldn't immediately figure out what part of the code was responsibe

cc @ElPicador